Исправил замену конфига пои его изменении

+++ Launcher/src/main/java/ru/gravit/launcher/ClientLauncherWrapper.java
@@ -22,27 +22,29 @@ public class ClientLauncherWrapper {
 
     public static void main(final String[] arguments) throws IOException, InterruptedException {
         final String pathLauncher = IOHelper.getCodeSource(ClientLauncher.class).toString();
-        System.out.println("Создаю новый файл конфигурации...");
-        FileWriter file = null;
-        JSONObject obj = new JSONObject();
-        obj.put("projectname", "Dreamfinity");
-        obj.put("address", "mc.dreamfinity.org");
-        obj.put("port", 7240);
-        obj.put("clientPort", 32288);
-        obj.put("secretKeyClient", "04e51b63006856bb");
-        obj.put("env", 3);
-        try {
-            file = new FileWriter(pathLauncher.replace(new File(ClientLauncherWrapper.class.getProtectionDomain().getCodeSource().getLocation().toString()).getName(), "")+"config.json");
-            file.write(obj.toJSONString());
-            System.out.println("Файл конфигурации создан!");
-        } catch (IOException ioException) {
-            ioException.printStackTrace();
-        } finally {
+        if (!new File(pathLauncher.replace(new File(ClientLauncherWrapper.class.getProtectionDomain().getCodeSource().getLocation().toString()).getName(), "")+"config.json").exists()) {
+            System.out.println("Создаю новый файл конфигурации...");
+            FileWriter file = null;
+            JSONObject obj = new JSONObject();
+            obj.put("projectname", "Dreamfinity");
+            obj.put("address", "mc.dreamfinity.org");
+            obj.put("port", 7240);
+            obj.put("clientPort", 32288);
+            obj.put("secretKeyClient", "04e51b63006856bb");
+            obj.put("env", 3);
             try {
-                file.flush();
-                file.close();
+                file = new FileWriter(pathLauncher.replace(new File(ClientLauncherWrapper.class.getProtectionDomain().getCodeSource().getLocation().toString()).getName(), "") + "config.json");
+                file.write(obj.toJSONString());
+                System.out.println("Файл конфигурации создан!");
             } catch (IOException ioException) {
                 ioException.printStackTrace();
+            } finally {
+                try {
+                    file.flush();
+                    file.close();
+                } catch (IOException ioException) {
+                    ioException.printStackTrace();
+                }
             }
         }
         LogHelper.printVersion("Launcher");
